refactor(data): rename lenses-g.json → lenses-gfx.json
Aligns the data filename with GFX branding used throughout the project
(URL path /lenses/gfx, diagram labels). Updates all import paths,
README, LICENSE-DATA, sitemap, and type comments.
