+++ LICENSE-DATA
@@ -1,6 +1,6 @@
 Data License — Proprietary
 
-Applies to: src/data/lenses.json
+Applies to: src/data/lenses.json, src/data/lenses-gfx.json
 Excludes: product images (src/data/images/) — all product images and brand trademarks
   are the property of their respective owners.
